+++ modules/System/admin.php
@@ -5,6 +5,7 @@ $this->helpers['settings'] = 'System\\Helper\\Settings';
 
 // Register routes
 $this->bindClass('System\\Controller\\Api', '/system/api');
+$this->bindClass('System\\Controller\\Buckets', '/system/buckets');
 $this->bindClass('System\\Controller\\Locales', '/system/locales');
 $this->bindClass('System\\Controller\\Logs', '/system/logs');
 $this->bindClass('System\\Controller\\Users\\Roles', '/system/users/roles');